fix integer problems

+++ application/boot/boot/model/member.py
@@ -3,8 +3,8 @@ from pweb_orm import PwebModel, pweb_orm
 
 class Member(PwebModel):
     name = pweb_orm.Column("name", pweb_orm.String(150), nullable=False)
-    mobile = pweb_orm.Column("mobile", pweb_orm.Integer(150), nullable=False)
-    emergencyContact = pweb_orm.Column("emergencyContact", pweb_orm.Integer(150), nullable=False)
+    mobile = pweb_orm.Column("mobile", pweb_orm.Integer(), nullable=False)
+    emergencyContact = pweb_orm.Column("emergencyContact", pweb_orm.Integer(), nullable=False)
     homeDistrict = pweb_orm.Column("homeDistrict", pweb_orm.String(150), nullable=False)
     studyIn = pweb_orm.Column("studyIn", pweb_orm.String(150), nullable=False)
     instituteName = pweb_orm.Column("instituteName", pweb_orm.String(150), nullable=False)
